+++ src/components/Sidebar.js
@@ -4,6 +4,7 @@ import Shares from "./Shares"
 import Likes from "./Likes"
 import Comments from "./Comments"
 
+
 const Sidebar = () => {
   return (
     <>
