+++ src/components/TopNav/Following.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import Tab from "@material-ui/core/Tab"
+import { Link } from 'react-router-dom'
+
+const Following = (props) => {
+    
+
+   return (
+        <>
+        <Tab label="Following" />
+        </>
+   )
+}
+
+export default Following
