+++ src/components/TopNav/Hashtags.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Tab from "@material-ui/core/Tab"
+
+const Hashtags = (props) => {
+   return (
+       <>
+       <Tab label="Hashtags" />
+       </>
+   )
+}
+
+export default Hashtags
